django-app/requirements.txt
django>=4.2,<5.0
opentelemetry-api>=1.20
opentelemetry-sdk>=1.20
opentelemetry-exporter-otlp-proto-grpc>=1.20
opentelemetry-instrumentation-django>=0.41b0
opentelemetry-instrumentation-requests>=0.41b0
requests>=2.31
python-json-logger>=2.0
